Recording counts for each handle separately.
Previously, we could define two different handles in such a way that
they would both be counted together. For example, if we defined
handle `:foo` and handle `:bar` both with the same interval, when we
incremented one of the counters the other one would be incremented
as well.

Now we keep the `:foo` counter and the `:bar` counter separate.