Add throttle method that does not raise exception
There are use cases where throttling is desired within a method without
an exception bubbling up. The combination of 'throttle!' and
'throttled?' can be used to largely get the effect, but an exception
will still be raised at the time the threshold is hit.

This will allow easy switching on whether or not the threshold has been
reached without the fear of an undesired exception popping up.